refactor(agent-tars): merge redundant setting constants across threads
